finalized

+++ js/carrito.js
@@ -1,4 +1,5 @@
-const licoresEnCarrito = JSON.parse(localStorage.getItem("mi-carrito"))
+let licoresEnCarrito = localStorage.getItem("mi-carrito") 
+licoresEnCarrito = JSON.parse(localStorage.getItem("mi-carrito"))
 
 
 const carritoLicores = document.querySelector("#carrito-licores")
@@ -21,8 +22,7 @@ function cargarLicoresAlCarrito(){
         carritoComprado.classList.add("disabled")
     
         carritoLicores.innerHTML = ""
-    
-    
+
         licoresEnCarrito.forEach(licor => {
     
             const div = document.createElement("div")
@@ -41,12 +41,12 @@ function cargarLicoresAlCarrito(){
     
                 <div class="carrito-licor-precio">
                     <h3>Precio</h3>
-                    <p>$${licor.precio}</p>
+                    <p>$ ${licor.precio}</p>
                 </div>
     
                 <div class="carrito-licor-subtotal">
                     <h3>Subtotal</h3>
-                    <p>$${licor.precio * licor.cantidad}</p>
+                    <p>$ ${licor.precio * licor.cantidad}</p>
                 </div>
                 <button class="carrito-licor-eliminar" id="${licor.codigo}"><i class="bi bi-trash3-fill"></i></button>
             `
@@ -55,6 +55,8 @@ function cargarLicoresAlCarrito(){
     
         });
         
+        actBotonEliminar()
+        actualizarTotal()
     
     }else{
     
@@ -64,9 +66,6 @@ function cargarLicoresAlCarrito(){
         carritoComprado.classList.add("disabled")
 
     }
-    
-    actBotonEliminar()
-    actualizarTotal()
 
 }
 
@@ -86,20 +85,59 @@ function eliminarDelCarrito(e){
     const idBoton = e.currentTarget.id
     
     const index = licoresEnCarrito.findIndex(licor => licor.codigo === parseInt(idBoton))
+
+    const licorEliminado = licoresEnCarrito[index].nombre
+
     licoresEnCarrito.splice(index, 1)
     cargarLicoresAlCarrito()
 
-
     localStorage.setItem("mi-carrito", JSON.stringify(licoresEnCarrito))
+
+    Toastify({
+        text: `Se eliminó ${licorEliminado} del carrito.`,
+        duration: 3000,
+        gravity: "bottom",
+        position: "right",
+        stopOnFocus: true,
+        style: {
+            background: "linear-gradient(to right, #ddbd48, #96c93d)",
+            borderRadius: '2rem',
+            fontSize: '1.2rem'
+        },
+        offset: {
+            x: '1.5rem', 
+            y: '5rem' 
+        },
+        onClick: function(){} // Callback after click
+    }).showToast();
 }
 
 botonVaciar.addEventListener("click", vaciarElCarrito)
 
+
 function vaciarElCarrito(){
 
-    licoresEnCarrito.length = 0
-    localStorage.setItem("mi-carrito", JSON.stringify(licoresEnCarrito))
-    cargarLicoresAlCarrito()
+    Swal.fire({
+        title: "¿Estás seguro?",
+        text: "Se van a eliminar todos los productos",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Si, eliminar",
+        cancelButtonText: "Cancelar"
+    }).then((result) => {
+        if (result.isConfirmed) {
+            licoresEnCarrito.length = 0
+            localStorage.setItem("mi-carrito", JSON.stringify(licoresEnCarrito))
+            cargarLicoresAlCarrito()
+            Swal.fire({
+                title: "Eliminado",
+                text: "Tus productos fueron eliminados correctamente",
+                icon: "success"
+            });
+        }
+    });
 
 }
 
@@ -107,18 +145,38 @@ function actualizarTotal(){
 
     const totalCalculado = licoresEnCarrito.reduce((acc, licor) => acc + (licor.precio * licor.cantidad), 0)
     totalCarrito.innerText = `$${totalCalculado}`
-
 }
 
 botonComprar.addEventListener("click", comprarCarrito)
 
 function comprarCarrito(){
-
-    licoresEnCarrito.length = 0
-    localStorage.setItem("mi-carrito", JSON.stringify(licoresEnCarrito))
     
-        carritoVacio.classList.add("disabled")
-        carritoLicores.classList.add("disabled")
-        carritoAcciones.classList.add("disabled")
-        carritoComprado.classList.remove("disabled")
+    Swal.fire({
+        title: "¿Quieres finalizar tu compra?",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Si",
+        cancelButtonText: "No"
+    }).then((result) => {
+        if (result.isConfirmed) {
+            Swal.fire({
+                position: "center",
+                icon: "success",
+                title: "Compra finalizada con éxito",
+                showConfirmButton: false,
+                timer: 1500
+            });
+
+
+            licoresEnCarrito.length = 0
+            localStorage.setItem("mi-carrito", JSON.stringify(licoresEnCarrito))
+            
+                carritoVacio.classList.add("disabled")
+                carritoLicores.classList.add("disabled")
+                carritoAcciones.classList.add("disabled")
+                carritoComprado.classList.remove("disabled")
+        };
+    })
 }
